+++ src/transactions/authenticator/account.ts
@@ -8,6 +8,7 @@ import { AnyPublicKey } from "../../core/crypto/anyPublicKey";
 import { AnySignature } from "../../core/crypto/anySignature";
 import { Ed25519PublicKey, Ed25519Signature } from "../../core/crypto/ed25519";
 import { MultiEd25519PublicKey, MultiEd25519Signature } from "../../core/crypto/multiEd25519";
+import { MultiKey } from "../../core/crypto/multiKey";
 import { AccountAuthenticatorVariant } from "../../types";
 
 export abstract class AccountAuthenticator extends Serializable {
@@ -22,6 +23,8 @@ export abstract class AccountAuthenticator extends Serializable {
         return AccountAuthenticatorMultiEd25519.load(deserializer);
       case AccountAuthenticatorVariant.SingleKey:
         return AccountAuthenticatorSingleKey.load(deserializer);
+      case AccountAuthenticatorVariant.MultiKey:
+        return AccountAuthenticatorMultiKey.load(deserializer);
       default:
         throw new Error(`Unknown variant index for AccountAuthenticator: ${index}`);
     }
@@ -118,5 +121,41 @@ export class AccountAuthenticatorSingleKey extends AccountAuthenticator {
     const public_key = AnyPublicKey.deserialize(deserializer);
     const signature = AnySignature.deserialize(deserializer);
     return new AccountAuthenticatorSingleKey(public_key, signature);
+  }
+}
+
+/**
+ * AccountAuthenticatorMultiKey for a multi signer
+ *
+ * @param public_keys MultiKey
+ * @param signatures Signature
+ *
+ */
+export class AccountAuthenticatorMultiKey extends AccountAuthenticator {
+  public readonly public_keys: MultiKey;
+
+  public readonly signatures: Array<AnySignature>;
+
+  public readonly signatures_bitmap: Uint8Array;
+
+  constructor(public_keys: MultiKey, signatures: Array<AnySignature>, signatures_bitmap: Uint8Array) {
+    super();
+    this.public_keys = public_keys;
+    this.signatures = signatures;
+    this.signatures_bitmap = signatures_bitmap;
+  }
+
+  serialize(serializer: Serializer): void {
+    serializer.serializeU32AsUleb128(AccountAuthenticatorVariant.MultiKey);
+    this.public_keys.serialize(serializer);
+    serializer.serializeVector<AnySignature>(this.signatures);
+    serializer.serializeBytes(this.signatures_bitmap);
+  }
+
+  static load(deserializer: Deserializer): AccountAuthenticatorMultiKey {
+    const public_keys = MultiKey.deserialize(deserializer);
+    const signatures = deserializer.deserializeVector(AnySignature);
+    const signatures_bitmap = deserializer.deserializeBytes();
+    return new AccountAuthenticatorMultiKey(public_keys, signatures, signatures_bitmap);
   }
 }